requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
langgraph
langsmith
streamlit
faiss-cpu
pypdf
pymupdf4llm
llama-cpp-python
huggingface-hub
requests
python-dotenv
pytest
accelerate
pydantic
duckduckgo-search
trafilatura
beautifulsoup4